backend/requirements.txt
fastapi
uvicorn[standard]
ormar[sqlite]
strawberry-graphql[fastapi]
fastapi-jwt-auth